2nd assignment ver.1.2.24

+++ 1st_assignment_main.py
@@ -73,7 +73,7 @@ class myCar(object):
             while True:
                 line = self.read_digit()
                 if time.time() - start > 2.5:
-                    self.move_front(100 * (end - start))
+                    self.move_front(int(100 * (end - start)))
                 if numpy.sum(line) == 5:
                     break
                 dot = numpy.dot(weight, line)
